still in progress changes to add winner at end of game
