+++ cycle/game/scripting/draw_actors_action.py
@@ -26,7 +26,7 @@ class DrawActorsAction(Action):
             cast (Cast): The cast of Actors in the game.
             script (Script): The script of Actions in the game.
         """
-        score = cast.get_first_actor("scores")
+        winner = cast.get_first_actor("winner")
 
         cycles = cast.get_actors("cycles")
         cycle1 = cycles[0]
@@ -38,6 +38,6 @@ class DrawActorsAction(Action):
         self._video_service.clear_buffer()
         self._video_service.draw_actors(cycle1_segments)
         self._video_service.draw_actors(cycle2_segments)
-        self._video_service.draw_actor(score)
+        self._video_service.draw_actor(winner)
         self._video_service.draw_actors(messages, True)
         self._video_service.flush_buffer()
